+++ cf/mixin/propertiesdatabounds.py
@@ -381,26 +381,36 @@ class PropertiesDataBounds(PropertiesData):
     **Bounds**
 
     How to operate on bounds is determined as follows, where "Flag" is
-    the boolean value returned by
+    the current value returned by
     ``cf.combine_bounds_with_coordinates()``
 
-    =====  ======  ======  ==========  =====================
-    Flag   x has   y has   z = x + y   Notes
-           bounds  bounds  has bounds
-    =====  ======  ======  ==========  =====================
-    False  Yes     Yes     Yes         `z.bounds` is
-                                       `x.bounds + y.bounds`
-    False  Yes     No      No
-    False  No      Yes     No
-    False  No      No      No
-    True   Yes     Yes     Yes         `z.bounds` is
-                                       `x.bounds + y.bounds`
-    True   Yes     No      Yes         `z.bounds` is
-                                       `x.bounds + y`
-    True   No      Yes     Yes         `z.bounds` is
-                                       `x + y.bounds`
-    True   No      No      No
-    =====  ======  ======  ==========  =====================
+    =======  ======  ======  ==========  =====================
+    Flag     x has   y has   z = x + y   Notes
+             bounds  bounds  has bounds
+    =======  ======  ======  ==========  =====================
+    `'AND'`  Yes     Yes     Yes         `z.bounds` is
+    `'AND'`                              `x.bounds + y.bounds`
+    `'AND'`  Yes     No      No
+    `'AND'`  No      Yes     No
+    `'AND'`  No      No      No
+    `'OR'`   Yes     Yes     Yes         `z.bounds` is
+                                         `x.bounds + y.bounds`
+    `'OR'`   Yes     No      Yes         `z.bounds` is
+                                         `x.bounds + y`
+    `'OR'`   No      Yes     Yes         `z.bounds` is
+                                         `x + y.bounds`
+    `'OR'`   No      No      No
+    `'XOR'`  Yes     Yes     No
+    `'XOR'`  Yes     No      Yes         `z.bounds` is
+                                         `x.bounds + y`
+    `'XOR'`  No      Yes     Yes         `z.bounds` is
+                                         `x + y.bounds`
+    `'XOR'`  No      No      No
+    `False`  Yes     Yes     No
+    `False`  Yes     No      No
+    `False`  No      Yes     No
+    `False`  No      No      No
+    =======  ======  ======  ==========  =====================
 
     If the *bounds* parameter is False then the result will have no
     bounds regardless of the value of
@@ -428,8 +438,10 @@ class PropertiesDataBounds(PropertiesData):
 
         '''
         inplace = (method[2] == 'i')
-
-        combine_bounds = bounds and combine_bounds_with_coordinates()
+        
+        bounds_OR = (bounds and combine_bounds_with_coordinates() == 'OR')
+        bounds_AND = (bounds and combine_bounds_with_coordinates() == 'AND')
+        bounds_XOR = (bounds and combine_bounds_with_coordinates() == 'XOR')
 
         has_bounds = self.has_bounds()
 
@@ -441,7 +453,10 @@ class PropertiesDataBounds(PropertiesData):
         except AttributeError:
             other_bounds = None
 
-        if combine_bounds and not has_bounds and other_bounds is not None:
+        if (
+                (bounds_OR or bounds_XOR)
+                and not has_bounds and other_bounds is not None
+        ):
             # --------------------------------------------------------
             # If self has no bounds but other does, then copy self for
             # use in constructing new bounds.
@@ -450,33 +465,41 @@ class PropertiesDataBounds(PropertiesData):
 
         new = super()._binary_operation(other, method)
 
-        if not bounds:
+        if not (bounds_OR or bounds_AND or bounds_XOR):
             # --------------------------------------------------------
             # Remove any bounds from the result
             # --------------------------------------------------------
             new.del_bounds(None)
 
         elif has_bounds and other_bounds is not None:
-            # --------------------------------------------------------
-            # Combine bounds that exist on both self and other
-            # --------------------------------------------------------
-            new_bounds = self.bounds._binary_operation(
-                other_bounds,
-                method)
+            if bounds_OR or bounds_AND:
+                # ----------------------------------------------------
+                # Combine bounds that exist on both self and other
+                # ----------------------------------------------------
+                new_bounds = self.bounds._binary_operation(
+                    other_bounds,
+                    method)
+                
+                if not inplace:
+                    new.set_bounds(new_bounds, copy=False)
 
-            if not inplace:
-                new.set_bounds(new_bounds, copy=False)
+            elif bounds_XOR:
+                # ----------------------------------------------------
+                # Both self and other have bounds, so remove the
+                # bounds from the result
+                # ----------------------------------------------------
+                new.del_bounds(None)
 
-        elif not combine_bounds:
+        elif bounds_AND:
             # --------------------------------------------------------
-            # Only one of self and other has bounds, so remove the
+            # At most one of self and other has bounds, so remove the
             # bounds from the result.
             # --------------------------------------------------------
             new.del_bounds(None)
 
         elif has_bounds:
             # --------------------------------------------------------
-            # other has no bounds: Combine self bounds with other
+            # other has no bounds, combine self bounds with other
             # coordinates
             # --------------------------------------------------------
             if numpy_size(other) > 1:
@@ -494,8 +517,8 @@ class PropertiesDataBounds(PropertiesData):
 
         elif other_bounds is not None:
             # --------------------------------------------------------
-            # self has no bounds: Combine self coordinates with other
-            # bounds
+            # self has no bounds, so combine self coordinates with
+            # other bounds
             # --------------------------------------------------------
             new_bounds = self._Bounds(data=original_self.data, copy=True)
             for i in range(other_bounds.ndim - other.ndim):
@@ -506,11 +529,11 @@ class PropertiesDataBounds(PropertiesData):
                 # fowl of the broadcasting rules (e.g. "ValueError:
                 # non-broadcastable output operand with shape (3,1)
                 # doesn't match the broadcast shape (3,2)")
-                methodo = method.replace('__i', '__', 1)
+                method2 = method.replace('__i', '__', 1)
             else:
-                methodo = method
+                method2 = method
 
-            new_bounds = new_bounds._binary_operation(other_bounds, methodo)
+            new_bounds = new_bounds._binary_operation(other_bounds, method2)
             new.set_bounds(new_bounds, copy=False)
 
         new._custom['direction'] = None